Modelos concluídos, DTOs concluídos

+++ src/main/java/com/marketplace/dto/ItemPedidoDTO.java
@@ -2,18 +2,19 @@ package com.marketplace.dto;
 
 import com.marketplace.model.ItemPedido;
 
+import java.math.BigDecimal;
+
 public record ItemPedidoDTO(
         String nomeProduto,
         int quantidade,
-        double precoUnitario,
-        double subTotal
+        BigDecimal precoUnitario,
+        BigDecimal subTotal
 ) {
     public ItemPedidoDTO(ItemPedido item) {
         this(
                 item.getProduto().getNome(),
                 item.getQuantidade(),
                 item.getPrecoUnitario(),
-                // Usa o método transiente da sua model ou calcula na hora
                 item.getValorTotal()
         );
     }
